requirements-simple.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0
python-jose==3.3.0
python-multipart==0.0.6
SQLAlchemy==1.4.47
